pyproject.toml
[project]
name = "markdown-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "jupyter>=1.1.1",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-openai>=0.3.11",
    "langgraph>=0.3.21",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "notebook>=7.3.3",
    "pymupdf>=1.25.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
